requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-redis
drf-yasg[swagger]
dj-database-url
django-environ